+++ shuup/admin/shop_provider.py
@@ -16,11 +16,19 @@ SHOP_SESSION_KEY = "admin_shop"
 
 
 class AdminShopProvider(object):
-
     def get_shop(self, request):
         if not request.user.is_staff:
             return None
 
+        cached_shop = getattr(request, "_cached_admin_shop", None)
+        if cached_shop:
+            return request._cached_admin_shop
+
+        shop = self._get_shop(request)
+        request._cached_admin_shop = shop
+        return shop
+
+    def _get_shop(self, request):
         # take the first if multishop is disabled
         if not settings.SHUUP_ENABLE_MULTIPLE_SHOPS:
             return Shop.objects.first()
@@ -55,6 +63,7 @@ class AdminShopProvider(object):
             # only can set if the user is superuser or is the shop staff
             if shop.staff_members.filter(pk=request.user.pk).exists() or request.user.is_superuser:
                 request.session[SHOP_SESSION_KEY] = shop.id
+                request._cached_admin_shop = shop
             else:
                 raise PermissionDenied(_("You must have the Access to Admin Panel permissions to this shop."))
 
@@ -62,6 +71,7 @@ class AdminShopProvider(object):
             self.unset_shop(request)
 
     def unset_shop(self, request):
+        request._cached_admin_shop = None
         if SHOP_SESSION_KEY in request.session:
             del request.session[SHOP_SESSION_KEY]